Join URL components in a better way
The better way is obviously "nested string manipulation functions" rather than a library as it was before. People clearly never join URL components in the real world, URLs are born the way they are and nobody should change them. Or it's just string
concantenation when calling requests methods, done over and over again...